# source/maas-toolbox/docs/swagger.yaml
basePath: /api/v1
definitions:
  api.AddGroupRequest:
    description: Request body for adding a group to a tier
    properties:
      group:
        description: Kubernetes group name to add
        example: premium-users
        type: string
    required:
    - group
    type: object
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.Tier:
    description: Tier configuration that maps Kubernetes groups to a subscription
      tier
    properties:
      description:
        description: Tier description
        example: Free tier for basic users
        type: string
      groups:
        description: List of Kubernetes groups
        example:
        - system:authenticated
        items:
          type: string
        type: array
      level:
        description: Tier level (non-negative integer)
        example: 1
        type: integer
      name:
        description: Tier name (immutable after creation)
        example: free
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/opendatahub-io/maas-billing
  description: |-
    REST API service for managing tier-to-group mappings in the Open Data Hub Model as a Service (MaaS) project.
    This API provides CRUD operations for managing tiers that map Kubernetes groups to user-defined subscription tiers.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Tier-to-Group Admin API
  version: "1.0"
paths:
  /tiers:
    get:
      description: Retrieve a list of all tiers in the system
      produces:
      - application/json
      responses:
        "200":
          description: List of tiers
          schema:
            items:
              $ref: '#/definitions/models.Tier'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List all tiers
      tags:
      - tiers
    post:
      consumes:
      - application/json
      description: Create a new tier with name, description, level, and groups. The
        tier name must be unique and cannot be changed after creation.
      parameters:
      - description: Tier object
        in: body
        name: tier
        required: true
        schema:
          $ref: '#/definitions/models.Tier'
      produces:
      - application/json
      responses:
        "201":
          description: Tier created successfully
          schema:
            $ref: '#/definitions/models.Tier'
        "400":
          description: Bad request - validation error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict - tier already exists
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Create a new tier
      tags:
      - tiers
  /tiers/{name}:
    delete:
      description: Delete a tier by its name
      parameters:
      - description: Tier name
        in: path
        name: name
        required: true
        type: string
      responses:
        "204":
          description: No content - tier deleted successfully
        "404":
          description: Tier not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Delete a tier
      tags:
      - tiers
    get:
      description: Retrieve a tier by its name
      parameters:
      - description: Tier name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tier details
          schema:
            $ref: '#/definitions/models.Tier'
        "404":
          description: Tier not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get a specific tier
      tags:
      - tiers
    put:
      consumes:
      - application/json
      description: Update a tier's description, level, or groups. The tier name cannot
        be changed.
      parameters:
      - description: Tier name
        in: path
        name: name
        required: true
        type: string
      - description: Tier update object (name field is ignored)
        in: body
        name: updates
        required: true
        schema:
          $ref: '#/definitions/models.Tier'
      produces:
      - application/json
      responses:
        "200":
          description: Updated tier
          schema:
            $ref: '#/definitions/models.Tier'
        "400":
          description: Bad request - validation error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Tier not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Update a tier
      tags:
      - tiers
  /tiers/{name}/groups:
    post:
      consumes:
      - application/json
      description: Add a Kubernetes group to a tier. The group must not already exist
        in the tier.
      parameters:
      - description: Tier name
        in: path
        name: name
        required: true
        type: string
      - description: Group to add
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/api.AddGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated tier with new group
          schema:
            $ref: '#/definitions/models.Tier'
        "400":
          description: Bad request - validation error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Tier not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict - group already exists
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Add a group to a tier
      tags:
      - groups
  /tiers/{name}/groups/{group}:
    delete:
      description: Remove a Kubernetes group from a tier
      parameters:
      - description: Tier name
        in: path
        name: name
        required: true
        type: string
      - description: Group name to remove
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Updated tier with group removed
          schema:
            $ref: '#/definitions/models.Tier'
        "404":
          description: Tier or group not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Remove a group from a tier
      tags:
      - groups
schemes:
- http
- https
swagger: "2.0"